requirements.txt
alembic==1.8.1
asgiref==3.5.2
astroid==2.3.3
attrs==22.1.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
cassandra-driver==3.25.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.3
cryptography==37.0.4
Django==4.0.6
django-crontab==0.7.1
django-quick-jobs==0.2.0
djangorestframework==3.14.0
Flask==2.0.3
Flask-Migrate==3.1.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.5.1
geomet==0.2.1.post1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
lazy-object-proxy==1.4.3
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
mccabe==0.6.1
packaging==21.3
pika==1.3.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pylint==2.4.4
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.20.0
pytz==2019.3
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
sqlparse==0.3.0
structlog==22.1.0
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.11.2
zipp==3.8.1
